refactored the TextInterface into a new file
